Final Version of Post Method
Restructure the code in main.py and models.py files.
Added missing parametters.
Imported uuid library.
Added database.py, services.py and selectors.py files for future usage.
